# openapi.yaml
openapi: 3.0.3
info:
  title: Football League Scoreboard API
  description: |
    Backend API expected by the frontend (`frontend/src/api/client.ts`).

    **Authentication:** Admin-only write operations require a Bearer token obtained
    from `POST /auth/login`. Public read endpoints do not require authentication.

    **Scoring:** Goal events on live matches automatically update match scores.
    Standings are calculated from finished matches (win = 3 pts, draw = 1 pt, loss = 0 pts;
    ranked by points, goal difference, goals for).
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local development

tags:
  - name: Auth
  - name: League
  - name: Teams
  - name: Players
  - name: Matches
  - name: Match Events
  - name: Standings

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Admin login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Admin logout
      operationId: logout
      responses:
        '204':
          description: Logged out successfully

  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated admin
      operationId: getCurrentUser
      security:
        - bearerAuth: []
        - {}
      description: |
        Returns the current admin when a valid token is provided.
        Returns `null` body or 204 when not authenticated (frontend treats null as logged out).
      responses:
        '200':
          description: Current user or null if not authenticated
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthUser'
                  - type: 'null'

  /league:
    get:
      tags: [League]
      summary: Get league information
      operationId: getLeague
      security: []
      responses:
        '200':
          description: League details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'

    patch:
      tags: [League]
      summary: Update league settings
      operationId: updateLeague
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeagueInput'
      responses:
        '200':
          description: Updated league
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /teams:
    get:
      tags: [Teams]
      summary: List all teams
      operationId: getTeams
      security: []
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'

    post:
      tags: [Teams]
      summary: Create a team
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamInput'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /teams/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    get:
      tags: [Teams]
      summary: Get a team by ID
      operationId: getTeam
      security: []
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Teams]
      summary: Update a team
      operationId: updateTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamInput'
      responses:
        '200':
          description: Updated team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Teams]
      summary: Delete a team
      operationId: deleteTeam
      description: |
        Fails if the team has assigned players or is referenced in any match.
      responses:
        '204':
          description: Team deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /players:
    get:
      tags: [Players]
      summary: List players
      operationId: getPlayers
      security: []
      parameters:
        - name: teamId
          in: query
          required: false
          description: Filter players by team
          schema:
            type: string
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'

    post:
      tags: [Players]
      summary: Create a player
      operationId: createPlayer
      description: |
        Assigns the player to a team. If `isCaptain` is true, any existing captain
        on that team is demoted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayerInput'
      responses:
        '201':
          description: Player created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /players/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    get:
      tags: [Players]
      summary: Get a player by ID
      operationId: getPlayer
      security: []
      responses:
        '200':
          description: Player details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Players]
      summary: Update a player
      operationId: updatePlayer
      description: |
        Supports reassigning a player to another team. If `isCaptain` is set to true,
        any existing captain on the target team is demoted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlayerInput'
      responses:
        '200':
          description: Updated player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Players]
      summary: Delete a player
      operationId: deletePlayer
      description: Fails if the player is referenced in any match event.
      responses:
        '204':
          description: Player deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches:
    get:
      tags: [Matches]
      summary: List matches
      operationId: getMatches
      security: []
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by match status
          schema:
            $ref: '#/components/schemas/MatchStatus'
      responses:
        '200':
          description: |
            Matches sorted by scheduled date/time descending (most recent first).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'

    post:
      tags: [Matches]
      summary: Schedule a new match
      operationId: createMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMatchInput'
      responses:
        '201':
          description: Match created with status `scheduled`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    get:
      tags: [Matches]
      summary: Get a match by ID
      operationId: getMatch
      security: []
      responses:
        '200':
          description: Match details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Matches]
      summary: Update a scheduled match
      operationId: updateMatch
      description: Only matches with status `scheduled` can be updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMatchInput'
      responses:
        '200':
          description: Updated match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches/{id}/start:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    post:
      tags: [Matches]
      summary: Start a match
      operationId: startMatch
      description: Transitions status from `scheduled` to `live`.
      responses:
        '200':
          description: Match started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    post:
      tags: [Matches]
      summary: Cancel a match
      operationId: cancelMatch
      description: |
        Transitions status to `cancelled`, resets scores to 0, and removes all events.
        Cannot cancel a `finished` match.
      responses:
        '200':
          description: Match cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches/{id}/finish:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    post:
      tags: [Matches]
      summary: Finish a match
      operationId: finishMatch
      description: Transitions status from `live` to `finished`. Final score counts toward standings.
      responses:
        '200':
          description: Match finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches/{id}/events:
    parameters:
      - $ref: '#/components/parameters/ResourceId'

    get:
      tags: [Match Events]
      summary: List events for a match
      operationId: getMatchEvents
      security: []
      responses:
        '200':
          description: Events sorted by minute ascending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchEvent'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Match Events]
      summary: Add an event to a live match
      operationId: addMatchEvent
      description: |
        Events can only be added while the match status is `live`.
        A `goal` event automatically increments the corresponding team's score.
        Substitutions require `playerInId` and `playerOutId`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventInput'
      responses:
        '201':
          description: Event created (match score updated if event type is goal)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /matches/{matchId}/events/{eventId}:
    parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
      - name: eventId
        in: path
        required: true
        schema:
          type: string

    delete:
      tags: [Match Events]
      summary: Remove an event from a live match
      operationId: removeMatchEvent
      description: |
        Events can only be removed while the match status is `live`.
        Removing a `goal` event automatically recalculates the match score.
      responses:
        '204':
          description: Event removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /standings:
    get:
      tags: [Standings]
      summary: Get league standings
      operationId: getStandings
      security: []
      description: |
        Calculated from finished matches. Ranked by points, then goal difference,
        then goals for.
      responses:
        '200':
          description: Standings table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StandingRow'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token returned by POST /auth/login

  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Team not found

    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@league.com
        password:
          type: string
          format: password
          example: admin123

    AuthUser:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    LoginResponse:
      type: object
      required: [email, token]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
          description: Bearer token for authenticated requests

    LeagueStatus:
      type: string
      enum: [active, inactive, completed]

    League:
      type: object
      required: [id, name, season, status]
      properties:
        id:
          type: string
        name:
          type: string
        season:
          type: string
          example: 2025/26
        status:
          $ref: '#/components/schemas/LeagueStatus'

    UpdateLeagueInput:
      type: object
      properties:
        name:
          type: string
        season:
          type: string
        status:
          $ref: '#/components/schemas/LeagueStatus'

    Team:
      type: object
      required: [id, name, logo, leagueId]
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: string
          description: Logo URL or emoji
        leagueId:
          type: string

    CreateTeamInput:
      type: object
      required: [name, logo]
      properties:
        name:
          type: string
        logo:
          type: string

    UpdateTeamInput:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string

    PlayerPosition:
      type: string
      enum: [GK, DEF, MID, FWD]

    Player:
      type: object
      required: [id, name, jerseyNumber, position, isCaptain, teamId]
      properties:
        id:
          type: string
        name:
          type: string
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
        position:
          $ref: '#/components/schemas/PlayerPosition'
        isCaptain:
          type: boolean
        teamId:
          type: string

    CreatePlayerInput:
      type: object
      required: [name, jerseyNumber, position, isCaptain, teamId]
      properties:
        name:
          type: string
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
        position:
          $ref: '#/components/schemas/PlayerPosition'
        isCaptain:
          type: boolean
        teamId:
          type: string

    UpdatePlayerInput:
      type: object
      properties:
        name:
          type: string
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
        position:
          $ref: '#/components/schemas/PlayerPosition'
        isCaptain:
          type: boolean
        teamId:
          type: string

    MatchStatus:
      type: string
      enum: [scheduled, live, finished, cancelled]

    Match:
      type: object
      required: [id, homeTeamId, awayTeamId, scheduledAt, status, homeScore, awayScore]
      properties:
        id:
          type: string
        homeTeamId:
          type: string
        awayTeamId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/MatchStatus'
        homeScore:
          type: integer
          minimum: 0
        awayScore:
          type: integer
          minimum: 0

    CreateMatchInput:
      type: object
      required: [homeTeamId, awayTeamId, scheduledAt]
      properties:
        homeTeamId:
          type: string
        awayTeamId:
          type: string
          description: Must differ from homeTeamId
        scheduledAt:
          type: string
          format: date-time

    UpdateMatchInput:
      type: object
      properties:
        homeTeamId:
          type: string
        awayTeamId:
          type: string
        scheduledAt:
          type: string
          format: date-time

    EventType:
      type: string
      enum: [goal, yellow_card, red_card, substitution]

    MatchEvent:
      type: object
      required: [id, matchId, teamId, playerId, eventType, minute, description]
      properties:
        id:
          type: string
        matchId:
          type: string
        teamId:
          type: string
        playerId:
          type: string
        eventType:
          $ref: '#/components/schemas/EventType'
        minute:
          type: integer
          minimum: 1
          maximum: 120
        description:
          type: string
        playerInId:
          type: string
          description: Required for substitution events
        playerOutId:
          type: string
          description: Required for substitution events

    CreateEventInput:
      type: object
      required: [teamId, playerId, eventType, minute]
      properties:
        teamId:
          type: string
        playerId:
          type: string
        eventType:
          $ref: '#/components/schemas/EventType'
        minute:
          type: integer
          minimum: 1
          maximum: 120
        description:
          type: string
          default: ''
        playerInId:
          type: string
          description: Required when eventType is substitution
        playerOutId:
          type: string
          description: Required when eventType is substitution

    StandingRow:
      type: object
      required:
        - position
        - teamId
        - teamName
        - played
        - wins
        - draws
        - losses
        - goalsFor
        - goalsAgainst
        - goalDifference
        - points
      properties:
        position:
          type: integer
          minimum: 1
        teamId:
          type: string
        teamName:
          type: string
        played:
          type: integer
          minimum: 0
        wins:
          type: integer
          minimum: 0
        draws:
          type: integer
          minimum: 0
        losses:
          type: integer
          minimum: 0
        goalsFor:
          type: integer
          minimum: 0
        goalsAgainst:
          type: integer
          minimum: 0
        goalDifference:
          type: integer
        points:
          type: integer
          minimum: 0

security:
  - bearerAuth: []